singleChallenge renders properly on mount
