+++ src/components/dashboard/challenges/ChallengeCard.tsx
@@ -1,4 +1,4 @@
-import { Button, Card, Text } from "@chakra-ui/react";
+import { Card, Text } from "@chakra-ui/react";
 import { Link } from "react-router-dom";
 import { Challenge } from "../../../utils/supabaseTypes";
